requirements.txt
# Core AI/ML Dependencies
qai_hub_models
executorch
torch>=2.0.0
torchvision>=0.15.0

# ONNX Runtime - QNN (Qualcomm Neural Network) Version
# IMPORTANT: Uninstall standard onnxruntime first:
#   pip uninstall onnxruntime onnxruntime-gpu
# Then install QNN version:
#   pip install onnxruntime-qnn
# Note: onnxruntime-qnn is for Qualcomm NPU devices
onnxruntime-qnn>=1.16.0


# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# API Framework
fastapi==0.115.5
uvicorn[standard]==0.34.0
pydantic==2.12.4
python-multipart==0.0.20

# HTTP Client for testing
requests>=2.31.0

# Hugging Face for model downloads
huggingface-hub>=0.20.0
